chatbot-rag/app/config.py
import os
from dotenv import load_dotenv

load_dotenv()

OPENAI_MODEL = os.getenv("OPENAI_MODEL", "gpt-4o-mini")
EMBEDDING_MODEL = os.getenv("EMBEDDING_MODEL", "text-embedding-3-small")

CHUNK_SIZE = int(os.getenv("CHUNK_SIZE", 500))
CHUNK_OVERLAP = int(os.getenv("CHUNK_OVERLAP", 50))

DOCS_PATH = os.getenv("DOCS_PATH", "data/docs")
VECTORSTORE_PATH = os.getenv("VECTORSTORE_PATH", "data/faiss_index")
