chatbot-rag/pyproject.toml
[project]
name = "chatbot-rag"
version = "0.1.0"
description = ""
authors = [
    { name = "Your Name", email = "you@example.com" }
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi==0.125.0",
    "langchain-community==0.4.1",
    "langchain-core==1.2.2",
    "langchain-openai==1.1.5",
    "langchain-text-splitters==1.1.0",
    "numpy==2.3.5",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "python-dotenv==1.2.1",
    "watchdog==6.0.0",
    "email-validator (>=2.3.0,<3.0.0)",
    "faiss-cpu (>=1.8)",
    "uvicorn (>=0.38.0,<0.39.0)"
]
[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "uvicorn (>=0.38.0,<0.39.0)"
]
